Chore Changed price ield from decimal to integer

+++ api/models.py
@@ -38,7 +38,7 @@ class User(AbstractUser):
 class Inventory(models.Model):
     name = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.IntegerField(max_digits=10,)
     quantity = models.IntegerField()
 
 class Transaction(models.Model):
